Refactor Iterable module: move IterableActionContext to a new file, update imports, and introduce IterableLogLevel enum
