+++ src/enums/IterableLogLevel.ts
@@ -1,0 +1,10 @@
+/**
+ * Enum representing what level of logs will Android and iOS project be printing on their consoles respectively.
+ */
+export enum IterableLogLevel {
+  debug = 1,
+  info = 2,
+  error = 3,
+}
+
+export default IterableLogLevel;
