+++ src/enums/index.ts
@@ -1,3 +1,4 @@
 export * from './IterableActionSource';
 export * from './IterableAuthResponseResult';
 export * from './IterableEventName';
+export * from './IterableLogLevel';
